Move read_areas code to dedicated file
